website/requirements.txt
Django==1.8.2
Markdown==2.6.6
Pillow==3.1.1
argparse==1.2.1
decorator==4.0.9
django-ckeditor==5.0.3
django-cors-headers==1.1.0
django-filter==0.13.0
django-grappelli==2.8.1
djangorestframework==3.3.3
ipython==4.1.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
simplegeneric==0.8.1
traitlets==4.1.0
wsgiref==0.1.2
